fix(8.5-2): server concurrency & type hardening
- TeamStringForIndex: panic → return "" + TS overload (AC1)
- OutcomeReason named Go type + constants; "natural" always emitted (AC2)
- UserRemovedHook lifecycle; emote.RemoveUser bounds lastEmoteAt map (AC3)
- D109 formally accepted with tracking comment (AC4)
- uint64 comparison fix in user handler + 32-bit build guard (AC5)
- D86/D105/D106/D109/D113/D116 annotated in deferred-work.md

+++ server/internal/user/handler.go
@@ -110,7 +110,7 @@ func (h *UserHandler) GetProfile(c echo.Context) error {
 		return apperr.ErrBadRequest
 	}
 
-	if uint(paramID) != authUserID {
+	if paramID != uint64(authUserID) {
 		return apperr.ErrForbidden
 	}
 
@@ -152,7 +152,7 @@ func (h *UserHandler) UpdatePreferences(c echo.Context) error {
 		return apperr.ErrBadRequest
 	}
 
-	if uint(paramID) != authUserID {
+	if paramID != uint64(authUserID) {
 		return apperr.ErrForbidden
 	}
 
@@ -195,7 +195,7 @@ func (h *UserHandler) ListMatches(c echo.Context) error {
 	if err != nil || paramID == 0 {
 		return apperr.ErrBadRequest
 	}
-	if uint(paramID) != authUserID {
+	if paramID != uint64(authUserID) {
 		return apperr.ErrForbidden
 	}
